pyproject.toml
[project]
name = "book-management-main"
version = "0.1.0"
description = ""
authors = [
    {name = "Demo",email = "hirenn158@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.122.0,<0.123.0)",
    "uvicorn[standard] (>=0.38.0,<0.39.0)",
    "sqlmodel (>=0.0.27,<0.0.28)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "alembic (>=1.17.2,<2.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "greenlet (>=3.2.4,<4.0.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "bcrypt (==4.0.1)",
    "google-generativeai (>=0.8.5,<0.9.0)",
    "chromadb (>=0.5.3,<0.6.0)",
    "sentence-transformers (>=2.7.0,<3.0.0)",
    "PyMuPDF (>=1.24.0,<2.0.0)",
    "wordninja (>=2.0.0,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.1,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "faker (>=38.2.0,<39.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)"
]
